+++ adapters/postgres/query.go
@@ -2,7 +2,8 @@ package postgres
 
 import (
 	"fmt"
-	"migoro/utils"
+
+	"github.com/brownhounds/migoro/utils"
 )
 
 func DatabaseExistsQuery() string {