+++ adapters/sqlite/sqlite.go
@@ -1,10 +1,11 @@
 package sqlite
 
 import (
-	"migoro/query"
-	"migoro/types"
-	"migoro/utils"
 	"os"
+
+	"github.com/brownhounds/migoro/query"
+	"github.com/brownhounds/migoro/types"
+	"github.com/brownhounds/migoro/utils"
 
 	"github.com/jmoiron/sqlx"
 	_ "github.com/mattn/go-sqlite3"