+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [Unreleased]
+
+### Changed
+
+- Package namespace from `migoro` to `github.com/brownhounds/migoro`
+
 ## [0.1.1] - 2024-07-10
 
 ### Added